requirements.txt
filelock==3.18.0
fsspec==2024.6.1
numpy==2.2.5
pandas==2.2.3
PyYAML==6.0.2
streamlit==1.45.0
sympy==1.13.1
torch==2.5.1+cu118
torchaudio==2.5.1+cu118
torchvision==0.20.1+cu118
ultralytics==8.3.128
ultralytics-thop==2.0.14